Adjust type

+++ src/io_win.c
@@ -29,7 +29,7 @@ ssize_t
 udx__recvmsg (udx_t *self, uv_buf_t *buf, struct sockaddr *addr) {
   DWORD bytes, flags = 0;
 
-  size_t addr_len = sizeof(*addr);
+  int addr_len = sizeof(*addr);
 
   int result = WSARecvFrom(
     self->handle.socket,
